feat(solana): S3 pass 2, enterprise transfer and transfer10x2 complete the family (#39)
The OPEN-1 decision finishes landing: enterprise transfer (2-in/2-out) and
transfer10x2 (10-in/2-out) join deposit, withdraw and disburse256, so the
Solana rail now carries the full enterprise family next to the consumer
four. Both ops follow the pass-1 pattern verbatim: arbiter key injected
from PoolConfig, receiver and authority ciphertexts riding inside the
verified public vector, one length-checked 1088 B kem ct, the shared
invariant gate and hardened marker-PDA path.

The brief's premise that the enterprise 10x2 would be the largest wire on
the rail turned out false once the worksheet was computed: enterprise ops
carry ONE kem ct because their receiver cts live in the publics, so
transfer10x2_priv with its two cts stays the tightest fit and the
enterprise 10x2 clears the v1 cap with headroom. The corrected claim is
gate-pinned and recorded in the spec, and the wire model now includes both
ComputeBudget instructions so the pinned numbers match a realistic
submission.

Family flags widened u8 to u16 (bit 8 was needed) by absorbing the
adjacent always-zero reserved byte, so existing config images read
identically; the future initialize and set_family_flags instructions must
write u16 LE, recorded in the spec. transfer10x2 becomes the rail's worst
CU op at ~492k measured, ~35% of the cap, budgets committed with ~5%
headroom.

+++ chains/solana/harness/tests/gate5_invariants.rs
@@ -39,7 +39,7 @@ fn expect_reject(env: &Env, error: PoolError) -> mollusk_svm::result::Instructio
 
 /// Swap the config image for one with `flags` — the consumer profile with a
 /// family switched off (the ModuleNotRegistered analogue, SOLR §2.1).
-fn set_family_flags(env: &mut Env, flags: u8) {
+fn set_family_flags(env: &mut Env, flags: u16) {
     let config_key = env.config_key;
     let data = bongtu_solana_harness::config_account_data(flags);
     let mut account = env
